fix: テキスト型PDF文字起こし処理を追加

+++ Kikagaku/ImageTranscription/views.py
@@ -1,8 +1,11 @@
 from django.shortcuts import render
+from django.conf import settings
 import numpy as np
 from .forms import ImageUploadForm, PDFUploadForm
 from .myocr import analyze_picture_bycv2, analyze_picture_bypillow
-from .processpdf import discern_pdf, pdfocr
+from .processpdf import discern_pdf, pdfocr, pdf_to_text
+import os
+import fitz
 
 # Create your views here.
 
@@ -57,19 +60,54 @@ def image(request):
 def pdf(request):
     if request.method == "POST":
         context = {}
+        error_message = None
         form = PDFUploadForm(request.POST, request.FILES)
-        print(form.is_valid())
         if form.is_valid():
             pdf = request.FILES['pdf']
             uploaded = form.save()
-            # discern = discern_pdf(uploaded.pdf.file)
-            pdfocr(uploaded.pdf.path, 2)
-            context = {
-                'form': form,
-                # 'error_message': discern
-            }
-            uploaded.delete()
-            return render(request, 'imagetranscription/pdf.html', context)
+            discern = discern_pdf(uploaded.pdf.file)
+            if discern == 'image':
+                page_number = 2
+                pdfocr(uploaded.pdf.path, page_number)
+                context = {
+                    'form': form,
+                }
+                uploaded.delete()
+                return render(request, 'imagetranscription/pdf.html', context)
+            else:
+                page_number = request.POST['page_number']
+                doc = fitz.open(uploaded.pdf.file)
+                total_pages = doc.page_count
+                try:
+                    int(page_number)
+                    if int(page_number) < 1 or int(page_number) > total_pages:
+                        error_message = f'指定するページ番号は 1 〜 {total_pages} の範囲で設定してください。'
+                except ValueError:
+                    error_message = 'ページ番号は数値で設定してください。'
+                else:
+                    page_number = int(page_number)
+
+                if error_message:
+                    context = {
+                        'form': form,
+                        'error_message': error_message,
+                    }
+                    
+                    return render(request, 'imagetranscription/pdf.html', context)
+                
+                page_number -= 1
+                
+                output_filename = f"page{page_number}.jpg"
+                output_path = os.path.join(settings.MEDIA_ROOT, 'pdfs', output_filename)
+                text, image_path = pdf_to_text(doc, output_path, page_number)
+                image_url = settings.MEDIA_URL + '/pdfs/' + output_filename
+                uploaded.delete()
+                context = {
+                    'form': form,
+                    'result_text': text,
+                    "image_url": image_url
+                }
+                return render(request, 'imagetranscription/pdf.html', context)
         else:
             context = {
                 'form': form,
